requirements.txt
appdirs==1.4.3
attrs==19.1.0
click==6.7
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
dataclasses==0.6
elevation==1.0.6
fasteners==0.15
Fiona==1.8.6
future==0.17.1
GDAL==2.4.0
geopandas==0.5.1
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==3.1.1
monotonic==1.5
munch==2.3.2
numpy==1.16.4
pandas==0.24.2
Pillow==6.1.0
pyparsing==2.4.0
pyproj==2.2.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
Shapely==1.6.4.post2
six==1.12.0
